refactor: code cleanups
