+++ packages/shared/src/graphql/posts.ts
@@ -77,7 +77,8 @@ export interface Ad {
 }
 
 export interface ParentComment {
-  username?: string;
+  handle?: string;
+  authorId?: string;
   authorName: string;
   authorImage: string;
   publishDate: Date | string;
